Move shape types to dedicated folder
