+++ client/src/game/shapes/aura.ts
@@ -5,7 +5,7 @@ import { computeVisibility } from "@/game/visibility/te/te";
 import { circleLineIntersection, xyEqual } from "@/game/visibility/te/triag";
 import { GlobalPoint, LocalPoint } from "../geom";
 import { g2l, g2lr, g2lz, getUnitDistance } from "../units";
-import { Circle } from "./circle";
+import { Circle } from "./variants/circle";
 import { Shape } from "./shape";
 
 export function drawAuras(shape: Shape, ctx: CanvasRenderingContext2D): void {
