+++ client/src/game/shapes/variants/boundingrect.ts
@@ -1,6 +1,6 @@
 import { GlobalPoint, Point, Ray, Vector } from "@/game/geom";
-import { g2lx, g2ly } from "../units";
-import { rotateAroundPoint } from "../utils";
+import { g2lx, g2ly } from "../../units";
+import { rotateAroundPoint } from "../../utils";
 
 export class BoundingRect {
     readonly w: number;
