+++ client/src/game/shapes/variants/circulartoken.ts
@@ -2,11 +2,11 @@ import { calcFontScale } from "@/core/utils";
 import { InitiativeData } from "@/game/comm/types/general";
 import { ServerCircularToken } from "@/game/comm/types/shapes";
 import { GlobalPoint } from "@/game/geom";
-import { Circle } from "@/game/shapes/circle";
+import { Circle } from "@/game/shapes/variants/circle";
 import { gameStore } from "@/game/store";
 import { g2l, g2lz } from "@/game/units";
 import * as tinycolor from "tinycolor2";
-import { SHAPE_TYPE } from "./types";
+import { SHAPE_TYPE } from "../types";
 
 export class CircularToken extends Circle {
     type: SHAPE_TYPE = "circulartoken";
