+++ client/src/game/shapes/variants/line.ts
@@ -1,10 +1,10 @@
 import { GlobalPoint } from "@/game/geom";
-import { BoundingRect } from "@/game/shapes/boundingrect";
 import { Shape } from "@/game/shapes/shape";
+import { BoundingRect } from "@/game/shapes/variants/boundingrect";
 import { g2l, g2lx, g2ly, g2lz } from "@/game/units";
-import { ServerLine } from "../comm/types/shapes";
-import { rotateAroundPoint } from "../utils";
-import { SHAPE_TYPE } from "./types";
+import { ServerLine } from "../../comm/types/shapes";
+import { rotateAroundPoint } from "../../utils";
+import { SHAPE_TYPE } from "../types";
 
 export class Line extends Shape {
     type: SHAPE_TYPE = "line";
@@ -44,6 +44,7 @@ export class Line extends Shape {
             line_width: this.lineWidth,
         });
     }
+
     get points(): number[][] {
         return [
             [...rotateAroundPoint(this.refPoint, this.center(), this.angle)],
