+++ client/src/game/shapes/variants/polygon.ts
@@ -1,10 +1,10 @@
-import { ServerPolygon } from "../comm/types/shapes";
-import { getDistanceToSegment, GlobalPoint } from "../geom";
-import { g2l, g2lx, g2ly, g2lz } from "../units";
-import { filterEqualPoints, getFogColour, getPointsCenter, rotateAroundPoint } from "../utils";
+import { ServerPolygon } from "../../comm/types/shapes";
+import { getDistanceToSegment, GlobalPoint } from "../../geom";
+import { g2l, g2lx, g2ly, g2lz } from "../../units";
+import { filterEqualPoints, getFogColour, getPointsCenter, rotateAroundPoint } from "../../utils";
+import { Shape } from "../shape";
+import { SHAPE_TYPE } from "../types";
 import { BoundingRect } from "./boundingrect";
-import { Shape } from "./shape";
-import { SHAPE_TYPE } from "./types";
 
 export class Polygon extends Shape {
     type: SHAPE_TYPE = "polygon";
