+++ client/src/game/shapes/variants/text.ts
@@ -1,9 +1,9 @@
 import { GlobalPoint } from "@/game/geom";
-import { BoundingRect } from "@/game/shapes/boundingrect";
 import { Shape } from "@/game/shapes/shape";
-import { ServerText } from "../comm/types/shapes";
-import { rotateAroundPoint } from "../utils";
-import { SHAPE_TYPE } from "./types";
+import { BoundingRect } from "@/game/shapes/variants/boundingrect";
+import { ServerText } from "../../comm/types/shapes";
+import { rotateAroundPoint } from "../../utils";
+import { SHAPE_TYPE } from "../types";
 
 export class Text extends Shape {
     type: SHAPE_TYPE = "text";
@@ -23,6 +23,7 @@ export class Text extends Shape {
         this.text = text;
         this.font = font;
     }
+
     asDict(): ServerText {
         return Object.assign(this.getBaseDict(), {
             text: this.text,
@@ -30,12 +31,15 @@ export class Text extends Shape {
             angle: this.angle,
         });
     }
+
     get points(): number[][] {
         return [[...rotateAroundPoint(this.refPoint, this.center(), this.angle)]];
     }
+
     getBoundingBox(): BoundingRect {
         return new BoundingRect(this.refPoint, 5, 5); // TODO: fix this bounding box
     }
+
     draw(ctx: CanvasRenderingContext2D): void {
         super.draw(ctx);
         ctx.font = this.font;
@@ -52,6 +56,7 @@ export class Text extends Shape {
         // ctx.restore();
         super.drawPost(ctx);
     }
+
     contains(_point: GlobalPoint): boolean {
         return false; // TODO
     }
