+++ client/src/game/ui/annotation.ts
@@ -2,8 +2,8 @@ import { InvalidationMode, SyncMode } from "@/core/comm/types";
 import { GlobalPoint, LocalPoint } from "@/game/geom";
 import { Layer } from "@/game/layers/layer";
 import { layerManager } from "@/game/layers/manager";
-import { Rect } from "@/game/shapes/rect";
-import { Text } from "@/game/shapes/text";
+import { Rect } from "@/game/shapes/variants/rect";
+import { Text } from "@/game/shapes/variants/text";
 import { gameStore } from "@/game/store";
 import { l2g } from "@/game/units";
 import { floorStore } from "../layers/store";
