+++ client/src/game/ui/tools/draw.vue
@@ -14,10 +14,10 @@ import { GlobalPoint, LocalPoint } from "@/game/geom";
 import { Layer } from "@/game/layers/layer";
 import { snapToPoint } from "@/game/layers/utils";
 import { layerManager } from "@/game/layers/manager";
-import { Circle } from "@/game/shapes/circle";
-import { Line } from "@/game/shapes/line";
-import { Polygon } from "@/game/shapes/polygon";
-import { Rect } from "@/game/shapes/rect";
+import { Circle } from "@/game/shapes/variants/circle";
+import { Line } from "@/game/shapes/variants/line";
+import { Polygon } from "@/game/shapes/variants/polygon";
+import { Rect } from "@/game/shapes/variants/rect";
 import { Shape } from "@/game/shapes/shape";
 import { gameStore } from "@/game/store";
 import { getUnitDistance, l2g, g2lx, g2ly, l2gz, clampGridLine } from "@/game/units";
