+++ client/src/game/ui/tools/ping.ts
@@ -1,7 +1,7 @@
 import { InvalidationMode, SyncMode } from "@/core/comm/types";
 import { GlobalPoint, LocalPoint } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
-import { Circle } from "@/game/shapes/circle";
+import { Circle } from "@/game/shapes/variants/circle";
 import { gameStore } from "@/game/store";
 import Tool from "@/game/ui/tools/tool.vue";
 import { l2g } from "@/game/units";
