+++ client/src/game/ui/tools/select.vue
@@ -8,7 +8,7 @@ import { EventBus } from "@/game/event-bus";
 import { GlobalPoint, LocalPoint, Ray, Vector } from "@/game/geom";
 import { Layer } from "@/game/layers/layer";
 import { snapToPoint } from "@/game/layers/utils";
-import { Rect } from "@/game/shapes/rect";
+import { Rect } from "@/game/shapes/variants/rect";
 import { gameStore } from "@/game/store";
 import { calculateDelta, ToolName, ToolFeatures } from "@/game/ui/tools/utils";
 import { g2l, g2lx, g2ly, l2g, l2gz } from "@/game/units";
@@ -17,10 +17,10 @@ import { visibilityStore } from "@/game/visibility/store";
 import { TriangulationTarget } from "@/game/visibility/te/pa";
 import { gameSettingsStore } from "../../settings";
 import { ToolBasics } from "./ToolBasics";
-import { Circle } from "../../shapes/circle";
-import { Line } from "../../shapes/line";
+import { Circle } from "../../shapes/variants/circle";
+import { Line } from "../../shapes/variants/line";
 import { SyncMode, InvalidationMode } from "../../../core/comm/types";
-import { BoundingRect } from "../../shapes/boundingrect";
+import { BoundingRect } from "../../shapes/variants/boundingrect";
 import { floorStore } from "@/game/layers/store";
 import { sendShapePositionUpdate, sendShapeSizeUpdate } from "../../api/emits/shape/core";
 import { Shape } from "@/game/shapes/shape";
