+++ client/tests/unit/game/visibility/te/pa.test.ts
@@ -1,11 +1,11 @@
 import { GlobalPoint } from "@/game/geom";
 import { Shape } from "@/game/shapes/shape";
-import { Polygon } from "@/game/shapes/polygon";
-import { Rect } from "@/game/shapes/rect";
+import { Polygon } from "@/game/shapes/variants/polygon";
+import { Rect } from "@/game/shapes/variants/rect";
 import { CDT } from "@/game/visibility/te/cdt";
 import { addShapesToTriag, deleteShapeFromTriag, TriangulationTarget, setCDT } from "@/game/visibility/te/pa";
 import { rotateAroundOrigin, xySmaller } from "@/game/visibility/te/triag";
-import { BaseRect } from "../../../../../src/game/shapes/baserect";
+import { BaseRect } from "../../../../../src/game/shapes/variants/baserect";
 
 jest.mock("@/game/api/socket", () => ({
     socket: {
